+++ src/app/server/rfi-pub-sub-engine.interface.ts
@@ -16,16 +16,30 @@ export interface PubEngine extends PubSubEngine {
 
 export type RfiPubSubEngine = PubEngine & PubSubEngine;
 
-export interface Payload {
+/**
+ * @note `d` is optional because not all payloads coming in are guaranteed to have it
+ * until we release this code.
+ * @chore to make `d` required: https://www.pivotaltracker.com/story/show/173474911
+ */
+export interface Payload<Discriminator = string> {
   marshalled_trace?: string;
-  // the discriminator
-  d?: string;
+  discriminator?: Discriminator;
 }
-export interface NodeChangePayload extends Payload {
-  d?: 'node-change';
+export interface NodeChangePayload extends Payload<'node-change'> {
   publisher_version: string;
   oid: string;
   id: string;
   action: string;
   deltas: ModelDelta[];
+}
+
+export interface SubscriptionCommandPayload extends Payload<'sub-command'> {
+  command: SubscriptionCommand;
+  google_app_version: string;
+}
+
+export enum SubscriptionCommand {
+  SUBSCRIPTION_START_VERSION = 'SUBSCRIPTION_START_VERSION',
+  SUBSCRIPTION_STOP_VERSION = 'SUBSCRIPTION_STOP_VERSION',
+  SUBSCRIPTION_RUN_VERSION = 'SUBSCRIPTION_RUN_VERSION'
 }